requirements.txt
aiogram<3.0
environs~=8.0.0
openpyxl
python-docx